add v2pe utils and supprot v2pe for fintune

+++ internvl_chat/internvl/patch/pad_data_collator.py
@@ -72,7 +72,7 @@ def concat_pad_data_collator(features, max_item_length=None, pad_id=0):
         feat['attention_mask'] = feat['input_ids'].ne(pad_id)
 
         if 'position_ids' in feat:
-            temp_position_ids = torch.LongTensor([pad_id] * max_item_length)
+            temp_position_ids = [pad_id] * max_item_length
             temp_position_ids[:feat['position_ids'].shape[0]] = feat['position_ids']
             feat['position_ids'] = temp_position_ids
